Finalisation du traitement pour récupérer le planning pour un jour et l'afficher .

+++ src/main/java/com/example/propotypage_et_interface_utilisateur/CasePourLeJourController.java
@@ -57,7 +57,7 @@ public class CasePourLeJourController implements Initializable {
     public void setOpacity(double value){
         this.anchorPane.setOpacity(value);
     }
-    public void setHeigth(int valueADiviser){
+    public void setHeigth(double valueADiviser){
         this.anchorPane.setPrefHeight(this.anchorPane.getPrefHeight()/valueADiviser);
     }
     @Override
